Added static friction (planning to switch implementation)
This was added to fix an important problem with the base implementation, and it didn't work. I'm starting to think the way to do this is to actually simulate each chain piece rather than take shortcuts. it would also probably simpler too since i wouldn't have to create 4 separate data structures and add extra data to EVERY environmental object